@@ -2,21 +2,21 @@
 import os
 from typing import Dict, Union, List
 
-from setuptools import setup, find_packages # type: ignore[import]
+from setuptools import setup, find_packages  # type: ignore[import]
 
 source_root = os.path.join(os.path.dirname(os.path.abspath(__file__)), "x3dh")
 
 version_scope: Dict[str, Dict[str, str]] = {}
-with open(os.path.join(source_root, "version.py")) as f:
+with open(os.path.join(source_root, "version.py"), encoding="utf-8") as f:
     exec(f.read(), version_scope)
 version = version_scope["__version__"]
 
 project_scope: Dict[str, Dict[str, Union[str, List[str]]]] = {}
-with open(os.path.join(source_root, "project.py")) as f:
+with open(os.path.join(source_root, "project.py"), encoding="utf-8") as f:
     exec(f.read(), project_scope)
 project = project_scope["project"]
 
-with open("README.md") as f:
+with open("README.md", encoding="utf-8") as f:
     long_description = f.read()
 
 classifiers = [
@@ -27,7 +27,6 @@ classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
 
-    "Programming Language :: Python :: 3.6",
     "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
@@ -52,20 +51,19 @@ del project["categories"]
 del project["year"]
 
 setup(
-    version = version["short"],
-    long_description = long_description,
-    long_description_content_type = "text/markdown",
-    license = "MIT",
-    packages = find_packages(),
-    install_requires = [
+    version=version["short"],
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    license="MIT",
+    packages=find_packages(),
+    install_requires=[
         "XEdDSA>=1.0.0,<2",
-        "cryptography>=3.3.2,<37",
-        "libnacl>=1.6.1,<=2",
-        "packaging>=20.9,<22"
+        "cryptography>=3.3.2",
+        "pydantic>=1.7.4"
     ],
-    python_requires = ">=3.6,<4",
-    include_package_data = True,
-    zip_safe = False,
-    classifiers = classifiers,
+    python_requires=">=3.7",
+    include_package_data=True,
+    zip_safe=False,
+    classifiers=classifiers,
     **project
 )
