@@ -10,7 +10,6 @@
 # directories to sys.path here. If the directory is relative to the documentation root,
 # use os.path.abspath to make it absolute, like shown here.
 import os
-import re
 import sys
 
 this_file_path = os.path.dirname(os.path.abspath(__file__))
@@ -23,7 +22,7 @@ from project import project     as __project
 
 project   = __project["name"]
 author    = __project["author"]
-copyright = "{}, {}".format(__project["year"], __project["author"])
+copyright = f"{__project['year']}, {__project['author']}"
 
 # The short X.Y version
 version = __version["short"]
@@ -60,16 +59,31 @@ html_theme = "sphinx_rtd_theme"
 # "default.css" will overwrite the builtin "default.css".
 html_static_path = [ "_static" ]
 
-# -- Autodoc Member Skipping -------------------------------------------------------------
+# -- Autodoc Configuration ---------------------------------------------------------------
 
-private_name_regex = re.compile(r"^_\w+__")
+# The following two options seem to be ignored...
+autodoc_typehints = "description"
+autodoc_type_aliases = { type_alias: f"{type_alias}" for type_alias in {
+    "JSONObject"
+} }
+
 def autodoc_skip_member_handler(app, what, name, obj, skip, options):
-    """
-    A very simple handler for the autodoc-skip-member event that skips everything
-    "private", aka starting with double underscores. Everything else is left untouched.
-    """
+    # Skip private members, i.e. those that start with double underscores but do not end in underscores
+    if name.startswith("__") and not name.endswith("_"):
+        return True
 
-    if private_name_regex.match(name):
+    # Could be achieved using exclude-members, but this is more comfy
+    if name in {
+        "__abstractmethods__",
+        "__dict__",
+        "__module__",
+        "__new__",
+        "__weakref__",
+        "_abc_impl"
+    }: return True
+
+    # Skip __init__s without documentation. Those are just used for type hints.
+    if name == "__init__" and obj.__doc__ is None:
         return True
 
     return None
