@@ -1,30 +1,48 @@
-# pylint: disable=broad-except
-# pylint: disable=too-many-nested-blocks
-# pylint: disable=too-many-statements
-
 import base64
 import json
 import os
 import random
 import time
-from typing import Dict, Optional, Any, Union, Type, List, Iterator
+from typing import Any, Dict, Iterator, List, Optional, Type, Union
 from unittest import mock
-import warnings
 
-import pytest # type: ignore[import]
 import x3dh
 
-# All test coroutines will be treated as marked.
-pytestmark = pytest.mark.asyncio
+
+__all__ = [  # pylint: disable=unused-variable
+    "test_configuration",
+    "test_key_agreements",
+    "test_migrations",
+    "test_old_signed_pre_key",
+    "test_pre_key_availability",
+    "test_pre_key_refill",
+    "test_serialization",
+    "test_signed_pre_key_rotation",
+    "test_signed_pre_key_signature_verification"
+]
+
 
 def flip_random_bit(data: bytes, exclude_msb: bool = False) -> bytes:
-    # For Curve25519, the most significant bit of the public key is always cleared/ignored, as per RFC 7748
-    # (on page 7). Thus, a bit flip of that bit does not make the signature verification fail, because the bit
-    # flip is ignored. The `exclude_msb` parameter can be used to disallow the bit flip to appear on the most
-    # significant bit and should be set when working with Curve25519 public keys.
+    """
+    Flip a random bit in a byte array.
+
+    Args:
+        data: The byte array to flip a random bit in.
+        exclude_msb: Whether the most significant bit of the byte array should be excluded from the random
+            selection. See note below.
+
+    For Curve25519, the most significant bit of the public key is always cleared/ignored, as per RFC 7748 (on
+    page 7). Thus, a bit flip of that bit does not make the signature verification fail, because the bit flip
+    is ignored. The `exclude_msb` parameter can be used to disallow the bit flip to appear on the most
+    significant bit and should be set when working with Curve25519 public keys.
+
+    Returns:
+        The data with a random bit flipped.
+    """
+
     while True:
         modify_byte = random.randrange(len(data))
-        modify_bit  = random.randrange(8)
+        modify_bit = random.randrange(8)
 
         # If the most significant bit was randomly chosen and `exclude_msb` is set, choose again.
         if not (exclude_msb and modify_byte == len(data) - 1 and modify_bit == 7):
@@ -34,32 +52,62 @@ def flip_random_bit(data: bytes, exclude_msb: bool = False) -> bytes:
     data_mut[modify_byte] ^= 1 << modify_bit
     return bytes(data_mut)
 
+
 bundles: Dict[bytes, x3dh.Bundle] = {}
 
+
 class ExampleState(x3dh.State):
-    async def _publish_bundle(self, bundle: x3dh.Bundle) -> Any:
-        bundles[bundle.ik] = bundle
+    """
+    A state implementation for testing, which simulates bundle uploads by storing them in a global variable,
+    and does some fancy public key encoding.
+    """
+
+    def _publish_bundle(self, bundle: x3dh.Bundle) -> None:
+        bundles[bundle.identity_key] = bundle
 
     @staticmethod
-    def _encode_public_key(curve: x3dh.Curve, key_type: x3dh.CurveType, pub: bytes) -> bytes:
-        curve_indicator: bytes = curve.value.encode("ASCII")
-        key_type_indicator: bytes = key_type.value.encode("ASCII")
+    def _encode_public_key(key_format: x3dh.IdentityKeyFormat, pub: bytes) -> bytes:
+        return b"\x42" + pub + b"\x13\x37" + key_format.value.encode("ASCII")
 
-        return curve_indicator + b"\x42" + pub + b"\x13\x37" + key_type_indicator
 
 def get_bundle(state: ExampleState) -> x3dh.Bundle:
-    if state.ik_mont in bundles:
-        return bundles[state.ik_mont]
-    if state.ik_ed in bundles:
-        return bundles[state.ik_ed]
+    """
+    Retrieve a bundle from the simulated server.
+
+    Args:
+        state: The state to retrieve the bundle for.
+
+    Returns:
+        The bundle.
+
+    Raises:
+        AssertionError: if the bundle was never "uploaded".
+    """
+
+    if state.bundle.identity_key in bundles:
+        return bundles[state.bundle.identity_key]
     assert False
 
-async def create_state(state_settings: Dict[str, Any]) -> ExampleState:
+
+def create_state(state_settings: Dict[str, Any]) -> ExampleState:
+    """
+    Create an :class:`ExampleState` and make sure the state creation worked as expected.
+
+    Args:
+        state_settings: Arguments to pass to :meth:`ExampleState.create`.
+
+    Returns:
+        The state.
+
+    Raises:
+        AssertionError: in case of failure.
+    """
+
     exc: Optional[BaseException] = None
     state: Optional[ExampleState] = None
     try:
-        state = await ExampleState.create(**state_settings)
-    except BaseException as e:
+        state = ExampleState.create(**state_settings)
+    except BaseException as e:  # pylint: disable=broad-except
         exc = e
     assert exc is None
     assert state is not None
@@ -67,16 +115,30 @@ async def create_state(state_settings: Dict[str, Any]) -> ExampleState:
 
     return state
 
-async def create_state_expect(
+
+def create_state_expect(
     state_settings: Dict[str, Any],
     expected_exception: Type[BaseException],
     expected_message: Union[str, List[str]]
 ) -> None:
+    """
+    Create an :class:`ExampleState`, but expect an exception to be raised during creation..
+
+    Args:
+        state_settings: Arguments to pass to :meth:`ExampleState.create`.
+        expected_exception: The exception type expected to be raised.
+        expected_message: The message expected to be raised, or a list of message snippets that should be part
+            of the exception message.
+
+    Raises:
+        AssertionError: in case of failure.
+    """
+
     exc: Optional[BaseException] = None
     state: Optional[ExampleState] = None
     try:
-        state = await ExampleState.create(**state_settings)
-    except BaseException as e:
+        state = ExampleState.create(**state_settings)
+    except BaseException as e:  # pylint: disable=broad-except
         exc = e
     assert state is None
 
@@ -86,57 +148,63 @@ async def create_state_expect(
     for expected_message_snippet in expected_message:
         assert expected_message_snippet in str(exc)
 
+
 def generate_settings(
-    info_string: str,
-    spk_timeout: int = 7,
-    opk_refill_threshold: int = 25,
-    opk_refill_target: int = 100,
-    all_possibilities: bool = False
+    info: bytes,
+    signed_pre_key_rotation_period: int = 7 * 24 * 60 * 60,
+    pre_key_refill_threshold: int = 25,
+    pre_key_refill_target: int = 100
 ) -> Iterator[Dict[str, Any]]:
-    for curve in [ x3dh.Curve.Curve448, x3dh.Curve.Curve25519 ]:
-        for internal_ik_type in [ x3dh.CurveType.Mont, x3dh.CurveType.Ed ]:
-            for external_ik_type in [ x3dh.CurveType.Mont, x3dh.CurveType.Ed ]:
-                for hash_function in [ x3dh.HashFunction.SHA_256, x3dh.HashFunction.SHA_512 ]:
-                    state_settings: Dict[str, Any] = {
-                        "curve": curve,
-                        "internal_ik_type": internal_ik_type,
-                        "external_ik_type": external_ik_type,
-                        "hash_function": hash_function,
-                        "info_string": info_string,
-                        "spk_timeout": spk_timeout,
-                        "opk_refill_threshold": opk_refill_threshold,
-                        "opk_refill_target": opk_refill_target
-                    }
+    """
+    Generate state creation arguments.
 
-                    if not all_possibilities:
-                        if internal_ik_type is x3dh.CurveType.Ed and external_ik_type is x3dh.CurveType.Mont:
-                            # This combination of internal and external identity key types is forbidden.
-                            continue
+    Args:
+        info: The info to use constantly.
+        signed_pre_key_rotation_period: The signed pre key rotation period to use constantly.
+        pre_key_refill_threshold: The pre key refill threshold to use constantly.
+        pre_key_refill_target. The pre key refill target to use constantly.
 
-                        if curve is x3dh.Curve.Curve448:
-                            # Curve448 is currently not fully supported.
-                            continue
+    Returns:
+        An iterator which yields a set of state creation arguments, returning all valid combinations of
+        identity key format and hash function with the given constant values.
+    """
 
-                    yield state_settings
+    for identity_key_format in [ x3dh.IdentityKeyFormat.CURVE_25519, x3dh.IdentityKeyFormat.ED_25519 ]:
+        for hash_function in [ x3dh.HashFunction.SHA_256, x3dh.HashFunction.SHA_512 ]:
+            state_settings: Dict[str, Any] = {
+                "identity_key_format": identity_key_format,
+                "hash_function": hash_function,
+                "info": info,
+                "signed_pre_key_rotation_period": signed_pre_key_rotation_period,
+                "pre_key_refill_threshold": pre_key_refill_threshold,
+                "pre_key_refill_target": pre_key_refill_target
+            }
 
-async def test_key_agreements() -> None:
-    for state_settings in generate_settings("test_key_agreements"):
-        state_a = await create_state(state_settings)
-        state_b = await create_state(state_settings)
+            yield state_settings
+
+
+def test_key_agreements() -> None:
+    """
+    Test the general key agreement functionality.
+    """
+
+    for state_settings in generate_settings("test_key_agreements".encode("ASCII")):
+        state_a = create_state(state_settings)
+        state_b = create_state(state_settings)
 
         # Store the current bundles
         bundle_a_before = get_bundle(state_a)
         bundle_b_before = get_bundle(state_b)
 
         # Perform the first, active half of the key agreement
-        shared_secret_active = await state_a.get_shared_secret_active(
+        shared_secret_active, associated_data_active, header = state_a.get_shared_secret_active(
             bundle_b_before,
             "ad appendix".encode("ASCII")
         )
 
         # Perform the second, passive half of the key agreement
-        shared_secret_passive = await state_b.get_shared_secret_passive(
-            shared_secret_active.header,
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(
+            header,
             "ad appendix".encode("ASCII")
         )
 
@@ -150,150 +218,154 @@ async def test_key_agreements() -> None:
         # The bundle of the passive party should have been modified and published again:
         assert bundle_b_after != bundle_b_before
 
-        # To be exact, only one one-time pre key should have been removed from the bundle:
-        assert bundle_b_after.ik        == bundle_b_before.ik
-        assert bundle_b_after.spk       == bundle_b_before.spk
-        assert bundle_b_after.spk_sig   == bundle_b_before.spk_sig
-        assert len(bundle_b_after.opks) == len(bundle_b_before.opks) - 1
-        assert all(opk in bundle_b_before.opks for opk in bundle_b_after.opks)
+        # To be exact, only one pre key should have been removed from the bundle:
+        assert bundle_b_after.identity_key == bundle_b_before.identity_key
+        assert bundle_b_after.signed_pre_key == bundle_b_before.signed_pre_key
+        assert bundle_b_after.signed_pre_key_sig == bundle_b_before.signed_pre_key_sig
+        assert len(bundle_b_after.pre_keys) == len(bundle_b_before.pre_keys) - 1
+        assert all(pre_key in bundle_b_before.pre_keys for pre_key in bundle_b_after.pre_keys)
 
         # Both parties should have derived the same shared secret and built the same
         # associated data:
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_active == associated_data_passive
 
         # It should not be possible to accept the same header again:
         try:
-            await state_b.get_shared_secret_passive(
-                shared_secret_active.header,
+            state_b.get_shared_secret_passive(
+                header,
                 "ad appendix".encode("ASCII")
             )
             assert False
-        except x3dh.KeyExchangeException as e:
-            assert "one-time pre key" in str(e)
-            assert "not available"    in str(e)
+        except x3dh.KeyAgreementException as e:
+            assert "pre key" in str(e)
+            assert "not available" in str(e)
 
-        # If the key agreement does not use a one-time pre key, it should be possible to accept the header
+        # If the key agreement does not use a pre key, it should be possible to accept the header
         # multiple times:
         bundle_b = get_bundle(state_b)
-        bundle_b = x3dh.Bundle(ik=bundle_b.ik, spk=bundle_b.spk, spk_sig=bundle_b.spk_sig, opks=[])
-
-        shared_secret_active = await state_a.get_shared_secret_active(bundle_b, require_opk=False)
-
-        shared_secret_passive = await state_b.get_shared_secret_passive(
-            shared_secret_active.header,
-            require_opk=False
+        bundle_b = x3dh.Bundle(
+            identity_key=bundle_b.identity_key,
+            signed_pre_key=bundle_b.signed_pre_key,
+            signed_pre_key_sig=bundle_b.signed_pre_key_sig,
+            pre_keys=set()
         )
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
 
-        shared_secret_passive = await state_b.get_shared_secret_passive(
-            shared_secret_active.header,
-            require_opk=False
+        shared_secret_active, associated_data_active, header = state_a.get_shared_secret_active(
+            bundle_b,
+            require_pre_key=False
         )
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
 
-async def test_configuration() -> None:
-    for state_settings in generate_settings("test_configuration", all_possibilities=True):
-        curve: x3dh.Curve = state_settings["curve"]
-        internal_ik_type: x3dh.CurveType = state_settings["internal_ik_type"]
-        external_ik_type: x3dh.CurveType = state_settings["external_ik_type"]
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(
+            header,
+            require_pre_key=False
+        )
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_active == associated_data_passive
 
-        if internal_ik_type is x3dh.CurveType.Ed and external_ik_type is x3dh.CurveType.Mont:
-            # This combination of internal and external identity key types is forbidden.
-            # The call should throw an exception without publishing the bundle.
-            await create_state_expect(state_settings, ValueError, [
-                "internal_ik_type",
-                "external_ik_type"
-            ])
-            continue
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(
+            header,
+            require_pre_key=False
+        )
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_active == associated_data_passive
 
 
-        if curve is x3dh.Curve.Curve448:
-            # Curve448 is currently not fully supported.
-            await create_state_expect(state_settings, NotImplementedError, "Curve448")
-            continue
+def test_configuration() -> None:
+    """
+    Test whether incorrect argument values are rejected correctly.
+    """
 
+    for state_settings in generate_settings("test_configuration".encode("ASCII")):
         # Before destorying the settings, make sure that the state could be created like that:
-        await create_state(state_settings)
+        create_state(state_settings)
 
-        # Pass in a non-ASCII info string:
-        state_settings["info_string"] = "😁"
-        await create_state_expect(state_settings, ValueError, "info_string")
-        state_settings["info_string"] = "test_configuration"
+        state_settings["info"] = "test_configuration".encode("ASCII")
 
         # Pass an invalid timeout for the signed pre key
-        state_settings["spk_timeout"] = 0
-        await create_state_expect(state_settings, ValueError, "spk_timeout")
-        state_settings["spk_timeout"] = -random.randrange(1, 2**64)
-        await create_state_expect(state_settings, ValueError, "spk_timeout")
-        state_settings["spk_timeout"] = 1
+        state_settings["signed_pre_key_rotation_period"] = 0
+        create_state_expect(state_settings, ValueError, "signed_pre_key_rotation_period")
+        state_settings["signed_pre_key_rotation_period"] = -random.randrange(1, 2**64)
+        create_state_expect(state_settings, ValueError, "signed_pre_key_rotation_period")
+        state_settings["signed_pre_key_rotation_period"] = 1
 
-        # Pass an invalid combination of opk_refill_threshold and opk_refill_target
-        # opk_refill_threshold too small
-        state_settings["opk_refill_threshold"] = 0
-        await create_state_expect(state_settings, ValueError, "opk_refill_threshold")
-        state_settings["opk_refill_threshold"] = 25
+        # Pass an invalid combination of pre_key_refill_threshold and pre_key_refill_target
+        # pre_key_refill_threshold too small
+        state_settings["pre_key_refill_threshold"] = 0
+        create_state_expect(state_settings, ValueError, "pre_key_refill_threshold")
+        state_settings["pre_key_refill_threshold"] = 25
 
-        # opk_refill_target too small
-        state_settings["opk_refill_target"] = 0
-        await create_state_expect(state_settings, ValueError, "opk_refill_target")
-        state_settings["opk_refill_target"] = 100
+        # pre_key_refill_target too small
+        state_settings["pre_key_refill_target"] = 0
+        create_state_expect(state_settings, ValueError, "pre_key_refill_target")
+        state_settings["pre_key_refill_target"] = 100
 
-        # opk_refill_threshold above opk_refill_target
-        state_settings["opk_refill_threshold"] = 100
-        state_settings["opk_refill_target"] = 25
-        await create_state_expect(state_settings, ValueError, [
-            "opk_refill_threshold",
-            "opk_refill_target"
+        # pre_key_refill_threshold above pre_key_refill_target
+        state_settings["pre_key_refill_threshold"] = 100
+        state_settings["pre_key_refill_target"] = 25
+        create_state_expect(state_settings, ValueError, [
+            "pre_key_refill_threshold",
+            "pre_key_refill_target"
         ])
-        state_settings["opk_refill_threshold"] = 25
-        state_settings["opk_refill_target"] = 100
+        state_settings["pre_key_refill_threshold"] = 25
+        state_settings["pre_key_refill_target"] = 100
 
-        # opk_refill_threshold equals opk_refill_target (this should succeed)
-        state_settings["opk_refill_threshold"] = 25
-        state_settings["opk_refill_target"] = 25
-        await create_state(state_settings)
-        state_settings["opk_refill_threshold"] = 25
-        state_settings["opk_refill_target"] = 100
+        # pre_key_refill_threshold equals pre_key_refill_target (this should succeed)
+        state_settings["pre_key_refill_threshold"] = 25
+        state_settings["pre_key_refill_target"] = 25
+        create_state(state_settings)
+        state_settings["pre_key_refill_threshold"] = 25
+        state_settings["pre_key_refill_target"] = 100
 
-async def test_opk_refill() -> None:
-    for state_settings in generate_settings("test_opk_refill", opk_refill_threshold=5, opk_refill_target=10):
-        state_a = await create_state(state_settings)
-        state_b = await create_state(state_settings)
 
-        # Verify that the bundle contains 100 one-time pre keys initially:
-        prev = len(get_bundle(state_b).opks)
-        assert prev == state_settings["opk_refill_target"]
+def test_pre_key_refill() -> None:
+    """
+    Test pre key refill.
+    """
+
+    for state_settings in generate_settings(
+        "test_pre_key_refill".encode("ASCII"),
+        pre_key_refill_threshold=5,
+        pre_key_refill_target=10
+    ):
+        state_a = create_state(state_settings)
+        state_b = create_state(state_settings)
+
+        # Verify that the bundle contains 100 pre keys initially:
+        prev = len(get_bundle(state_b).pre_keys)
+        assert prev == state_settings["pre_key_refill_target"]
 
         # Perform a lot of key agreements and verify that the refill works as expected:
         for _ in range(100):
-            shared_secret_active = await state_a.get_shared_secret_active(get_bundle(state_b))
-            await state_b.get_shared_secret_passive(shared_secret_active.header)
+            header = state_a.get_shared_secret_active(get_bundle(state_b))[2]
+            state_b.get_shared_secret_passive(header)
 
-            num_opks = len(get_bundle(state_b).opks)
+            num_pre_keys = len(get_bundle(state_b).pre_keys)
 
-            if prev == state_settings["opk_refill_threshold"]:
-                assert num_opks == state_settings["opk_refill_target"]
+            if prev == state_settings["pre_key_refill_threshold"]:
+                assert num_pre_keys == state_settings["pre_key_refill_target"]
             else:
-                assert num_opks == prev - 1
+                assert num_pre_keys == prev - 1
 
-            prev = num_opks
+            prev = num_pre_keys
 
-async def test_spk_signature_verification() -> None:
-    for state_settings in generate_settings("test_spk_signature_verification"):
-        curve: x3dh.Curve = state_settings["curve"]
-        external_ik_type: x3dh.CurveType = state_settings["external_ik_type"]
+
+def test_signed_pre_key_signature_verification() -> None:
+    """
+    Test signature verification of the signed pre key.
+    """
+
+    for state_settings in generate_settings("test_signed_pre_key_signature_verification".encode("ASCII")):
+        identity_key_format: x3dh.IdentityKeyFormat = state_settings["identity_key_format"]
 
         for _ in range(100):
-            state_a = await create_state(state_settings)
-            state_b = await create_state(state_settings)
+            state_a = create_state(state_settings)
+            state_b = create_state(state_settings)
 
             bundle = get_bundle(state_b)
 
             # First, make sure that the active half of the key agreement usually works:
-            await state_a.get_shared_secret_active(bundle)
+            state_a.get_shared_secret_active(bundle)
 
             # Now, flip a random bit in
             # 1. the signature
@@ -302,119 +374,164 @@ async def test_spk_signature_verification() -> None:
             # and make sure that the active half of the key agreement reject the signature.
 
             # 1.: the signature
-            spk_sig = flip_random_bit(bundle.spk_sig)
-            bundle_modified = x3dh.Bundle(ik=bundle.ik, spk=bundle.spk, spk_sig=spk_sig, opks=bundle.opks)
+            signed_pre_key_sig = flip_random_bit(bundle.signed_pre_key_sig)
+            bundle_modified = x3dh.Bundle(
+                identity_key=bundle.identity_key,
+                signed_pre_key=bundle.signed_pre_key,
+                signed_pre_key_sig=signed_pre_key_sig,
+                pre_keys=bundle.pre_keys
+            )
             try:
-                await state_a.get_shared_secret_active(bundle_modified)
+                state_a.get_shared_secret_active(bundle_modified)
                 assert False
-            except x3dh.KeyExchangeException as e:
+            except x3dh.KeyAgreementException as e:
                 assert "signature" in str(e)
 
             # 2.: the identity key
-            exclude_msb = curve is x3dh.Curve.Curve25519 and external_ik_type is x3dh.CurveType.Mont
-            ik = flip_random_bit(bundle.ik, exclude_msb=exclude_msb)
-            bundle_modified = x3dh.Bundle(ik=ik, spk=bundle.spk, spk_sig=bundle.spk_sig, opks=bundle.opks)
+            exclude_msb = identity_key_format is x3dh.IdentityKeyFormat.CURVE_25519
+            identity_key = flip_random_bit(bundle.identity_key, exclude_msb=exclude_msb)
+            bundle_modified = x3dh.Bundle(
+                identity_key=identity_key,
+                signed_pre_key=bundle.signed_pre_key,
+                signed_pre_key_sig=bundle.signed_pre_key_sig,
+                pre_keys=bundle.pre_keys
+            )
             try:
-                await state_a.get_shared_secret_active(bundle_modified)
+                state_a.get_shared_secret_active(bundle_modified)
                 assert False
-            except x3dh.KeyExchangeException as e:
+            except x3dh.KeyAgreementException as e:
                 assert "signature" in str(e)
 
             # 3.: the signed pre key
-            spk = flip_random_bit(bundle.spk)
-            bundle_modified = x3dh.Bundle(ik=bundle.ik, spk=spk, spk_sig=bundle.spk_sig, opks=bundle.opks)
+            signed_pre_key = flip_random_bit(bundle.signed_pre_key)
+            bundle_modified = x3dh.Bundle(
+                identity_key=bundle.identity_key,
+                signed_pre_key=signed_pre_key,
+                signed_pre_key_sig=bundle.signed_pre_key_sig,
+                pre_keys=bundle.pre_keys
+            )
             try:
-                await state_a.get_shared_secret_active(bundle_modified)
+                state_a.get_shared_secret_active(bundle_modified)
                 assert False
-            except x3dh.KeyExchangeException as e:
+            except x3dh.KeyAgreementException as e:
                 assert "signature" in str(e)
 
-async def test_opk_availability() -> None:
-    for state_settings in generate_settings("test_opk_availability"):
-        state_a = await create_state(state_settings)
-        state_b = await create_state(state_settings)
+
+def test_pre_key_availability() -> None:
+    """
+    Test whether key agreements without pre keys work/are rejected as expected.
+    """
+
+    for state_settings in generate_settings("test_pre_key_availability".encode("ASCII")):
+        state_a = create_state(state_settings)
+        state_b = create_state(state_settings)
 
         # First, test the active half of the key agreement
-        for require_opk in [ True, False ]:
-            for include_opk in [ True, False ]:
+        for require_pre_key in [ True, False ]:
+            for include_pre_key in [ True, False ]:
                 bundle = get_bundle(state_b)
 
-                # Make sure that the bundle contains one-time pre keys:
-                assert len(bundle.opks) > 0
+                # Make sure that the bundle contains pre keys:
+                assert len(bundle.pre_keys) > 0
 
-                # If required for the test, remove all one-time pre keys:
-                if not include_opk:
-                    bundle = x3dh.Bundle(ik=bundle.ik, spk=bundle.spk, spk_sig=bundle.spk_sig, opks=[])
-
-                should_fail = require_opk and not include_opk
-                try:
-                    shared_secret_active = await state_a.get_shared_secret_active(
-                        bundle,
-                        require_opk=require_opk
+                # If required for the test, remove all pre keys:
+                if not include_pre_key:
+                    bundle = x3dh.Bundle(
+                        identity_key=bundle.identity_key,
+                        signed_pre_key=bundle.signed_pre_key,
+                        signed_pre_key_sig=bundle.signed_pre_key_sig,
+                        pre_keys=set()
                     )
+
+                should_fail = require_pre_key and not include_pre_key
+                try:
+                    header = state_a.get_shared_secret_active(
+                        bundle,
+                        require_pre_key=require_pre_key
+                    )[2]
                     assert not should_fail
-                    assert (shared_secret_active.header.opk is not None) == include_opk
-                except x3dh.KeyExchangeException as e:
+                    assert (header.pre_key is not None) == include_pre_key
+                except x3dh.KeyAgreementException as e:
                     assert should_fail
                     assert "does not contain" in str(e)
-                    assert "one-time pre key" in str(e)
+                    assert "pre key" in str(e)
 
         # Second, test the passive half of the key agreement
-        for require_opk in [ True, False ]:
-            for include_opk in [ True, False ]:
+        for require_pre_key in [ True, False ]:
+            for include_pre_key in [ True, False ]:
                 bundle = get_bundle(state_b)
 
-                # Make sure that the bundle contains one-time pre keys:
-                assert len(bundle.opks) > 0
+                # Make sure that the bundle contains pre keys:
+                assert len(bundle.pre_keys) > 0
 
-                # If required for the test, remove all one-time pre keys:
-                if not include_opk:
-                    bundle = x3dh.Bundle(ik=bundle.ik, spk=bundle.spk, spk_sig=bundle.spk_sig, opks=[])
+                # If required for the test, remove all pre keys:
+                if not include_pre_key:
+                    bundle = x3dh.Bundle(
+                        identity_key=bundle.identity_key,
+                        signed_pre_key=bundle.signed_pre_key,
+                        signed_pre_key_sig=bundle.signed_pre_key_sig,
+                        pre_keys=set()
+                    )
 
-                # Perform the active half of the key agreement, using a one-time pre key only if required for
+                # Perform the active half of the key agreement, using a pre key only if required for
                 # the test.
-                shared_secret_active = await state_a.get_shared_secret_active(bundle, require_opk=False)
+                shared_secret_active, _, header = state_a.get_shared_secret_active(
+                    bundle,
+                    require_pre_key=False
+                )
 
-                should_fail = require_opk and not include_opk
+                should_fail = require_pre_key and not include_pre_key
                 try:
-                    shared_secret_passive = await state_b.get_shared_secret_passive(
-                        shared_secret_active.header,
-                        require_opk=require_opk
+                    shared_secret_passive, _, _ = state_b.get_shared_secret_passive(
+                        header,
+                        require_pre_key=require_pre_key
                     )
                     assert not should_fail
-                    assert shared_secret_passive.shared_secret == shared_secret_active.shared_secret
-                except x3dh.KeyExchangeException as e:
+                    assert shared_secret_passive == shared_secret_active
+                except x3dh.KeyAgreementException as e:
                     assert should_fail
                     assert "does not use" in str(e)
-                    assert "one-time pre key" in str(e)
+                    assert "pre key" in str(e)
+
 
 THREE_DAYS = 3 * 24 * 60 * 60
 EIGHT_DAYS = 8 * 24 * 60 * 60
 
-async def test_spk_rotation() -> None:
-    for state_settings in generate_settings("test_spk_rotation"):
-        state_b  = await create_state(state_settings)
+
+def test_signed_pre_key_rotation() -> None:
+    """
+    Test signed pre key rotation logic.
+    """
+
+    for state_settings in generate_settings("test_signed_pre_key_rotation".encode("ASCII")):
+        state_b = create_state(state_settings)
         bundle_b = get_bundle(state_b)
 
         current_time = time.time()
-        time_mock    = mock.MagicMock()
+        time_mock = mock.MagicMock()
 
         # Mock time.time, so that the test can skip days in an instant
         with mock.patch("time.time", time_mock):
             # ExampleState.create should call time.time only once, when generating the signed pre key. Make
             # the mock return the actual current time for that call.
             time_mock.return_value = current_time
-            state_a = await create_state(state_settings)
+            state_a = create_state(state_settings)
             assert time_mock.call_count == 1
             time_mock.reset_mock()
 
-            # Prepare a key agreement header, the time is irrelevant here. Don't use a one-time pre key so
+            # Prepare a key agreement header, the time is irrelevant here. Don't use a pre key so
             # that the header can be used multiple times.
             bundle_a = get_bundle(state_a)
-            bundle_a = x3dh.Bundle(ik=bundle_a.ik, spk=bundle_a.spk, spk_sig=bundle_a.spk_sig, opks=[])
+            bundle_a = x3dh.Bundle(
+                identity_key=bundle_a.identity_key,
+                signed_pre_key=bundle_a.signed_pre_key,
+                signed_pre_key_sig=bundle_a.signed_pre_key_sig,
+                pre_keys=set()
+            )
 
             time_mock.return_value = current_time + THREE_DAYS
-            shared_secret_b = await state_b.get_shared_secret_active(bundle_a, require_opk=False)
+            header_b = state_b.get_shared_secret_active(bundle_a, require_pre_key=False)[2]
+            state_b.rotate_signed_pre_key()
             assert time_mock.call_count == 1
             time_mock.reset_mock()
 
@@ -429,7 +546,8 @@ async def test_spk_rotation() -> None:
             # rotation.
             bundle_a_before = get_bundle(state_a)
             time_mock.return_value = current_time + THREE_DAYS
-            await state_a.get_shared_secret_active(bundle_b)
+            state_a.get_shared_secret_active(bundle_b)
+            state_a.rotate_signed_pre_key()
             assert time_mock.call_count == 1
             time_mock.reset_mock()
             assert get_bundle(state_a) == bundle_a_before
@@ -438,13 +556,14 @@ async def test_spk_rotation() -> None:
             # A rotation reads the time twice.
             bundle_a_before = get_bundle(state_a)
             time_mock.return_value = current_time + EIGHT_DAYS
-            await state_a.get_shared_secret_active(bundle_b)
+            state_a.get_shared_secret_active(bundle_b)
+            state_a.rotate_signed_pre_key()
             assert time_mock.call_count == 2
             time_mock.reset_mock()
-            assert get_bundle(state_a).ik      == bundle_a_before.ik
-            assert get_bundle(state_a).spk     != bundle_a_before.spk
-            assert get_bundle(state_a).spk_sig != bundle_a_before.spk_sig
-            assert get_bundle(state_a).opks    == bundle_a_before.opks
+            assert get_bundle(state_a).identity_key == bundle_a_before.identity_key
+            assert get_bundle(state_a).signed_pre_key != bundle_a_before.signed_pre_key
+            assert get_bundle(state_a).signed_pre_key_sig != bundle_a_before.signed_pre_key_sig
+            assert get_bundle(state_a).pre_keys == bundle_a_before.pre_keys
 
             # Update the "current_time" to the creation time of the last signed pre key:
             current_time += EIGHT_DAYS
@@ -455,7 +574,8 @@ async def test_spk_rotation() -> None:
             # rotation.
             bundle_a_before = get_bundle(state_a)
             time_mock.return_value = current_time + THREE_DAYS
-            await state_a.get_shared_secret_passive(shared_secret_b.header, require_opk=False)
+            state_a.get_shared_secret_passive(header_b, require_pre_key=False)
+            state_a.rotate_signed_pre_key()
             assert time_mock.call_count == 1
             time_mock.reset_mock()
             assert get_bundle(state_a) == bundle_a_before
@@ -464,13 +584,14 @@ async def test_spk_rotation() -> None:
             # A rotation reads the time twice.
             bundle_a_before = get_bundle(state_a)
             time_mock.return_value = current_time + EIGHT_DAYS
-            await state_a.get_shared_secret_passive(shared_secret_b.header, require_opk=False)
+            state_a.get_shared_secret_passive(header_b, require_pre_key=False)
+            state_a.rotate_signed_pre_key()
             assert time_mock.call_count == 2
             time_mock.reset_mock()
-            assert get_bundle(state_a).ik      == bundle_a_before.ik
-            assert get_bundle(state_a).spk     != bundle_a_before.spk
-            assert get_bundle(state_a).spk_sig != bundle_a_before.spk_sig
-            assert get_bundle(state_a).opks    == bundle_a_before.opks
+            assert get_bundle(state_a).identity_key == bundle_a_before.identity_key
+            assert get_bundle(state_a).signed_pre_key != bundle_a_before.signed_pre_key
+            assert get_bundle(state_a).signed_pre_key_sig != bundle_a_before.signed_pre_key_sig
+            assert get_bundle(state_a).pre_keys == bundle_a_before.pre_keys
 
             # Update the "current_time" to the creation time of the last signed pre key:
             current_time += EIGHT_DAYS
@@ -481,7 +602,7 @@ async def test_spk_rotation() -> None:
             # rotation.
             bundle_a_before = get_bundle(state_a)
             time_mock.return_value = current_time + THREE_DAYS
-            state_a = await ExampleState.deserialize(state_a.serialize(), **state_settings)
+            state_a = ExampleState.from_model(state_a.model, **state_settings)
             assert time_mock.call_count == 1
             time_mock.reset_mock()
             assert get_bundle(state_a) == bundle_a_before
@@ -490,255 +611,224 @@ async def test_spk_rotation() -> None:
             # A rotation reads the time twice.
             bundle_a_before = get_bundle(state_a)
             time_mock.return_value = current_time + EIGHT_DAYS
-            state_a = await ExampleState.deserialize(state_a.serialize(), **state_settings)
+            state_a = ExampleState.from_model(state_a.model, **state_settings)
             assert time_mock.call_count == 2
             time_mock.reset_mock()
-            assert get_bundle(state_a).ik      == bundle_a_before.ik
-            assert get_bundle(state_a).spk     != bundle_a_before.spk
-            assert get_bundle(state_a).spk_sig != bundle_a_before.spk_sig
-            assert get_bundle(state_a).opks    == bundle_a_before.opks
+            assert get_bundle(state_a).identity_key == bundle_a_before.identity_key
+            assert get_bundle(state_a).signed_pre_key != bundle_a_before.signed_pre_key
+            assert get_bundle(state_a).signed_pre_key_sig != bundle_a_before.signed_pre_key_sig
+            assert get_bundle(state_a).pre_keys == bundle_a_before.pre_keys
 
             # Update the "current_time" to the creation time of the last signed pre key:
             current_time += EIGHT_DAYS
 
-async def test_old_spk() -> None:
-    for state_settings in generate_settings("test_old_spk", spk_timeout=2):
-        state_a = await create_state(state_settings)
-        state_b = await create_state(state_settings)
 
-        # Prepare a key agreement header using the current signed pre key of state a. Don't use a one-time pre
+def test_old_signed_pre_key() -> None:
+    """
+    Test that the old signed pre key remains available for key agreements for one further rotation period.
+    """
+
+    for state_settings in generate_settings(
+        "test_old_signed_pre_key".encode("ASCII"),
+        signed_pre_key_rotation_period=2
+    ):
+        print(state_settings)
+        state_a = create_state(state_settings)
+        state_b = create_state(state_settings)
+
+        # Prepare a key agreement header using the current signed pre key of state a. Don't use a pre
         # key so that the header can be used multiple times.
         bundle_a = get_bundle(state_a)
-        bundle_a_no_opks = x3dh.Bundle(ik=bundle_a.ik, spk=bundle_a.spk, spk_sig=bundle_a.spk_sig, opks=[])
-        shared_secret_active = await state_b.get_shared_secret_active(bundle_a_no_opks, require_opk=False)
-        header = shared_secret_active.header
+        bundle_a_no_pre_keys = x3dh.Bundle(
+            identity_key=bundle_a.identity_key,
+            signed_pre_key=bundle_a.signed_pre_key,
+            signed_pre_key_sig=bundle_a.signed_pre_key_sig,
+            pre_keys=set()
+        )
+        shared_secret_active, associated_data_active, header = state_b.get_shared_secret_active(
+            bundle_a_no_pre_keys,
+            require_pre_key=False
+        )
 
         # Make sure that this key agreement works as intended:
-        shared_secret_passive = await state_a.get_shared_secret_passive(header, require_opk=False)
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        shared_secret_passive, associated_data_passive, _ = state_a.get_shared_secret_passive(
+            header,
+            require_pre_key=False
+        )
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_active == associated_data_passive
 
         # Rotate the signed pre key once. The rotation period is specified as two days, still skipping eight
         # days should only trigger a single rotation.
         current_time = time.time()
-        time_mock    = mock.MagicMock()
+        time_mock = mock.MagicMock()
 
         # Mock time.time, so that the test can skip days in an instant
         with mock.patch("time.time", time_mock):
             time_mock.return_value = current_time + EIGHT_DAYS
-            state_a = await ExampleState.deserialize(state_a.serialize(), **state_settings)
+            state_a = ExampleState.from_model(state_a.model, **state_settings)
             assert time_mock.call_count == 2
             time_mock.reset_mock()
 
         # Make sure that the signed pre key was rotated:
-        assert get_bundle(state_a).ik      == bundle_a.ik
-        assert get_bundle(state_a).spk     != bundle_a.spk
-        assert get_bundle(state_a).spk_sig != bundle_a.spk_sig
-        assert get_bundle(state_a).opks    == bundle_a.opks
+        assert get_bundle(state_a).identity_key == bundle_a.identity_key
+        assert get_bundle(state_a).signed_pre_key != bundle_a.signed_pre_key
+        assert get_bundle(state_a).signed_pre_key_sig != bundle_a.signed_pre_key_sig
+        assert get_bundle(state_a).pre_keys == bundle_a.pre_keys
 
         bundle_a_rotated = get_bundle(state_a)
 
         # The old signed pre key should still be stored in state_a, thus the old key agreement header should
         # still work:
-        shared_secret_passive = await state_a.get_shared_secret_passive(header, require_opk=False)
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        shared_secret_passive, associated_data_passive, _ = state_a.get_shared_secret_passive(
+            header,
+            require_pre_key=False
+        )
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_active == associated_data_passive
 
         # Rotate the signed pre key again:
         with mock.patch("time.time", time_mock):
             time_mock.return_value = current_time + EIGHT_DAYS + THREE_DAYS
-            state_a = await ExampleState.deserialize(state_a.serialize(), **state_settings)
+            state_a = ExampleState.from_model(state_a.model, **state_settings)
             assert time_mock.call_count == 2
             time_mock.reset_mock()
 
         # Make sure that the signed pre key was rotated again:
-        assert get_bundle(state_a).ik      == bundle_a.ik
-        assert get_bundle(state_a).spk     != bundle_a.spk
-        assert get_bundle(state_a).spk_sig != bundle_a.spk_sig
-        assert get_bundle(state_a).opks    == bundle_a.opks
-        assert get_bundle(state_a).ik      == bundle_a_rotated.ik
-        assert get_bundle(state_a).spk     != bundle_a_rotated.spk
-        assert get_bundle(state_a).spk_sig != bundle_a_rotated.spk_sig
-        assert get_bundle(state_a).opks    == bundle_a_rotated.opks
+        assert get_bundle(state_a).identity_key == bundle_a.identity_key
+        assert get_bundle(state_a).signed_pre_key != bundle_a.signed_pre_key
+        assert get_bundle(state_a).signed_pre_key_sig != bundle_a.signed_pre_key_sig
+        assert get_bundle(state_a).pre_keys == bundle_a.pre_keys
+        assert get_bundle(state_a).identity_key == bundle_a_rotated.identity_key
+        assert get_bundle(state_a).signed_pre_key != bundle_a_rotated.signed_pre_key
+        assert get_bundle(state_a).signed_pre_key_sig != bundle_a_rotated.signed_pre_key_sig
+        assert get_bundle(state_a).pre_keys == bundle_a_rotated.pre_keys
 
         # Now the signed pre key used in the header should not be available any more, the passive half of the
         # key agreement should fail:
         try:
-            await state_a.get_shared_secret_passive(header, require_opk=False)
+            state_a.get_shared_secret_passive(header, require_pre_key=False)
             assert False
-        except x3dh.KeyExchangeException as e:
+        except x3dh.KeyAgreementException as e:
             assert "signed pre key" in str(e)
-            assert "not available"  in str(e)
+            assert "not available" in str(e)
 
-async def test_serialization() -> None:
-    for state_settings in generate_settings("test_serialization"):
-        state_a = await create_state(state_settings)
-        state_b = await create_state(state_settings)
+
+def test_serialization() -> None:
+    """
+    Test (de)serialization.
+    """
+
+    for state_settings in generate_settings("test_serialization".encode("ASCII")):
+        state_a = create_state(state_settings)
+        state_b = create_state(state_settings)
 
         # Make sure that the key agreement works normally:
-        shared_secret_active  = await state_a.get_shared_secret_active(get_bundle(state_b))
-        shared_secret_passive = await state_b.get_shared_secret_passive(shared_secret_active.header)
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        shared_secret_active, associated_data_acitve, header = state_a.get_shared_secret_active(
+            get_bundle(state_b)
+        )
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(header)
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_acitve == associated_data_passive
 
         # Do the same thing but serialize and deserialize state b before performing the passive half of the
         # key agreement:
         bundle_b_before = get_bundle(state_b)
 
-        shared_secret_active = await state_a.get_shared_secret_active(get_bundle(state_b))
-        state_b = await ExampleState.deserialize(state_b.serialize(), **state_settings)
-        shared_secret_passive = await state_b.get_shared_secret_passive(shared_secret_active.header)
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        shared_secret_active, associated_data_acitve, header = state_a.get_shared_secret_active(
+            get_bundle(state_b)
+        )
+        state_b = ExampleState.from_model(state_b.model, **state_settings)
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(header)
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_acitve == associated_data_passive
 
-        # Make sure that the bundle remained the same, except for one one-time pre key being deleted:
-        assert get_bundle(state_b).ik        == bundle_b_before.ik
-        assert get_bundle(state_b).spk       == bundle_b_before.spk
-        assert get_bundle(state_b).spk_sig   == bundle_b_before.spk_sig
-        assert len(get_bundle(state_b).opks) == len(bundle_b_before.opks) - 1
-        assert all(opk in bundle_b_before.opks for opk in get_bundle(state_b).opks)
+        # Make sure that the bundle remained the same, except for one pre key being deleted:
+        assert get_bundle(state_b).identity_key == bundle_b_before.identity_key
+        assert get_bundle(state_b).signed_pre_key == bundle_b_before.signed_pre_key
+        assert get_bundle(state_b).signed_pre_key_sig == bundle_b_before.signed_pre_key_sig
+        assert len(get_bundle(state_b).pre_keys) == len(bundle_b_before.pre_keys) - 1
+        assert all(pre_key in bundle_b_before.pre_keys for pre_key in get_bundle(state_b).pre_keys)
 
-        # Accepting a key agreement using a one-time pre key results in the one-time pre key being deleted
-        # from the state. Use (de)serialization to circumvent the deletion of the one-time pre key. This time
+        # Accepting a key agreement using a pre key results in the pre key being deleted
+        # from the state. Use (de)serialization to circumvent the deletion of the pre key. This time
         # also serialize the structure into JSON:
-        shared_secret_active = await state_a.get_shared_secret_active(get_bundle(state_b))
-        state_b_serialized = json.dumps(state_b.serialize())
+        shared_secret_active, associated_data_acitve, header = state_a.get_shared_secret_active(
+            get_bundle(state_b)
+        )
+        state_b_serialized = json.dumps(state_b.json)
 
         # Accepting the header should work once...
-        shared_secret_passive = await state_b.get_shared_secret_passive(shared_secret_active.header)
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(header)
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_acitve == associated_data_passive
 
         # ...but fail the second time:
         try:
-            await state_b.get_shared_secret_passive(shared_secret_active.header)
+            state_b.get_shared_secret_passive(header)
             assert False
-        except x3dh.KeyExchangeException as e:
-            assert "one-time pre key" in str(e)
-            assert "not available"    in str(e)
+        except x3dh.KeyAgreementException as e:
+            assert "pre key" in str(e)
+            assert "not available" in str(e)
 
         # After restoring the state, it should work again:
-        state_b = await ExampleState.deserialize(json.loads(state_b_serialized), **state_settings)
-        shared_secret_passive = await state_b.get_shared_secret_passive(shared_secret_active.header)
-        assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-        assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+        state_b, needs_publish = ExampleState.from_json(json.loads(state_b_serialized), **state_settings)
+        shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(header)
+        assert not needs_publish
+        assert shared_secret_active == shared_secret_passive
+        assert associated_data_acitve == associated_data_passive
 
-async def test_serialization_consistency_checks() -> None:
-    # When deserializing, the code checks whether the settings/configuration has changed and generates
-    # warnings/errors accordingly.
-
-    state_settings: Dict[str, Any] = {
-        "curve": x3dh.Curve.Curve25519,
-        "internal_ik_type": x3dh.CurveType.Mont,
-        "external_ik_type": x3dh.CurveType.Ed,
-        "hash_function": x3dh.HashFunction.SHA_256,
-        "info_string": "test_serialization_consistency_checks",
-        "spk_timeout": 7,
-        "opk_refill_threshold": 25,
-        "opk_refill_target": 100
-    }
-
-    state = await create_state(state_settings)
-
-    # Serialize state a with the current settings
-    state_serialized = state.serialize()
-
-    # Make sure that the serialization generally works:
-    await ExampleState.deserialize(state_serialized, **state_settings)
-
-    # Modify the "curve" setting and verify that `deserialize` throws an exception:
-    state_settings["curve"] = x3dh.Curve.Curve448
-    try:
-        await ExampleState.deserialize(state_serialized, **state_settings)
-        assert False
-    except x3dh.InconsistentConfigurationException as e:
-        assert "uses keys on" in str(e)
-    state_settings["curve"] = x3dh.Curve.Curve25519
-
-    # Modify the "internal_ik_type" setting and verify that `deserialize` throws an exception:
-    state_settings["internal_ik_type"] = x3dh.CurveType.Ed
-    try:
-        await ExampleState.deserialize(state_serialized, **state_settings)
-        assert False
-    except x3dh.InconsistentConfigurationException as e:
-        assert "internal" in str(e)
-        assert "identity key" in str(e)
-    state_settings["internal_ik_type"] = x3dh.CurveType.Mont
-
-    # Modify the "external_ik_type" setting and verify that `deserialize` generates a warning:
-    state_settings["external_ik_type"] = x3dh.CurveType.Mont
-    with warnings.catch_warnings(record=True) as w:
-        await ExampleState.deserialize(state_serialized, **state_settings)
-        assert len(w) == 1
-        assert issubclass(w[0].category, UserWarning)
-        assert "external" in str(w[0].message)
-        assert "identity key" in str(w[0].message)
-    state_settings["external_ik_type"] = x3dh.CurveType.Ed
-
-    # Modify the "hash_function" setting and verify that `deserialize` generates a warning:
-    state_settings["hash_function"] = x3dh.HashFunction.SHA_512
-    with warnings.catch_warnings(record=True) as w:
-        await ExampleState.deserialize(state_serialized, **state_settings)
-        assert len(w) == 1
-        assert issubclass(w[0].category, UserWarning)
-        assert "hash function" in str(w[0].message)
-    state_settings["hash_function"] = x3dh.HashFunction.SHA_256
-
-    # Modify the "info_string" setting and verify that `deserialize` generates a warning:
-    state_settings["info_string"] = "something different than before"
-    with warnings.catch_warnings(record=True) as w:
-        await ExampleState.deserialize(state_serialized, **state_settings)
-        assert len(w) == 1
-        assert issubclass(w[0].category, UserWarning)
-        assert "info string" in str(w[0].message)
-    state_settings["info_string"] = "test_serialization_consistency_checks"
-
-    # Finally a sanity check: verify that all settings were restored and the deserialization works again:
-    await ExampleState.deserialize(state_serialized, **state_settings)
 
 THIS_FILE_PATH = os.path.dirname(os.path.abspath(__file__))
 
-async def test_migrations() -> None:
-    # Test the migration from pre-stable
+
+def test_migrations() -> None:
+    """
+    Test the migration from pre-stable.
+    """
+
     state_settings: Dict[str, Any] = {
-        "curve": x3dh.Curve.Curve25519,
-        "internal_ik_type": x3dh.CurveType.Mont,
-        "external_ik_type": x3dh.CurveType.Mont,
+        "identity_key_format": x3dh.IdentityKeyFormat.CURVE_25519,
         "hash_function": x3dh.HashFunction.SHA_256,
-        "info_string": "test_migrations",
-        "spk_timeout": 7,
-        "opk_refill_threshold": 25,
-        "opk_refill_target": 100
+        "info": "test_migrations".encode("ASCII"),
+        "signed_pre_key_rotation_period": 7,
+        "pre_key_refill_threshold": 25,
+        "pre_key_refill_target": 100
     }
 
-    with open(os.path.join(THIS_FILE_PATH, "migration_data", "state-alice-pre-stable.json"), "r") as f:
-        state_a_serialized = json.load(f)
+    with open(os.path.join(
+        THIS_FILE_PATH,
+        "migration_data",
+        "state-alice-pre-stable.json"
+    ), "r", encoding="utf-8") as state_alice_pre_stable_json:
+        state_a_serialized = json.load(state_alice_pre_stable_json)
 
-    with open(os.path.join(THIS_FILE_PATH, "migration_data", "state-bob-pre-stable.json"), "r") as f:
-        state_b_serialized = json.load(f)
+    with open(os.path.join(
+        THIS_FILE_PATH,
+        "migration_data",
+        "state-bob-pre-stable.json"
+    ), "r", encoding="utf-8") as state_bob_pre_stable_json:
+        state_b_serialized = json.load(state_bob_pre_stable_json)
 
-    with open(os.path.join(THIS_FILE_PATH, "migration_data", "shared-secret-pre-stable.json"), "r") as f:
-        shared_secret_active_serialized = json.load(f)
+    with open(os.path.join(
+        THIS_FILE_PATH,
+        "migration_data",
+        "shared-secret-pre-stable.json"
+    ), "r", encoding="utf-8") as shared_secret_pey_stable_json:
+        shared_secret_active_serialized = json.load(shared_secret_pey_stable_json)
 
     # Convert the pre-stable shared secret structure into a x3dh.SharedSecretActive
-    shared_secret_active = x3dh.SharedSecretActive(
-        shared_secret   = base64.b64decode(shared_secret_active_serialized["sk"].encode("ASCII")),
-        associated_data = base64.b64decode(shared_secret_active_serialized["ad"].encode("ASCII")),
-        header = x3dh.Header(
-            ik  = base64.b64decode(shared_secret_active_serialized["to_other"]["ik"].encode("ASCII")),
-            ek  = base64.b64decode(shared_secret_active_serialized["to_other"]["ek"].encode("ASCII")),
-            spk = base64.b64decode(shared_secret_active_serialized["to_other"]["spk"].encode("ASCII")),
-            opk = base64.b64decode(shared_secret_active_serialized["to_other"]["otpk"].encode("ASCII"))
-        )
+    shared_secret_active = base64.b64decode(shared_secret_active_serialized["sk"].encode("ASCII"))
+    associated_data_active = base64.b64decode(shared_secret_active_serialized["ad"].encode("ASCII"))
+    header = x3dh.Header(
+        identity_key=base64.b64decode(shared_secret_active_serialized["to_other"]["ik"].encode("ASCII")),
+        ephemeral_key=base64.b64decode(shared_secret_active_serialized["to_other"]["ek"].encode("ASCII")),
+        signed_pre_key=base64.b64decode(shared_secret_active_serialized["to_other"]["spk"].encode("ASCII")),
+        pre_key=base64.b64decode(shared_secret_active_serialized["to_other"]["otpk"].encode("ASCII"))
     )
 
     # Load state a. This should not trigger a publishing of the bundle, as the `changed` flag is not set.
-    # A warning will be printed due missing information in the pre-stable serialization format.
-    with warnings.catch_warnings(record=True) as w:
-        state_a = await ExampleState.deserialize(state_a_serialized, **state_settings)
-        assert len(w) == 1
-        assert issubclass(w[0].category, UserWarning)
-        assert "pre-stable" in str(w[0].message)
+    state_a, _needs_publish = ExampleState.from_json(state_a_serialized, **state_settings)
 
     try:
         get_bundle(state_a)
@@ -747,22 +837,19 @@ async def test_migrations() -> None:
         pass
 
     # Load state b. This should trigger a publishing of the bundle, as the `changed` flag is set.
-    # A warning will be printed due missing information in the pre-stable serialization format.
-    with warnings.catch_warnings(record=True) as w:
-        state_b = await ExampleState.deserialize(state_b_serialized, **state_settings)
-        assert len(w) == 1
-        assert issubclass(w[0].category, UserWarning)
-        assert "pre-stable" in str(w[0].message)
+    state_b, _needs_publish = ExampleState.from_json(state_b_serialized, **state_settings)
 
     get_bundle(state_b)
 
     # Complete the passive half of the key agreement as created by the pre-stable version:
-    shared_secret_passive = await state_b.get_shared_secret_passive(shared_secret_active.header)
-    assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-    assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+    shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(header)
+    assert shared_secret_active == shared_secret_passive
+    # Don't check the associated data, since formats have changed.
 
     # Try another key agreement using the migrated sessions:
-    shared_secret_active  = await state_a.get_shared_secret_active(get_bundle(state_b))
-    shared_secret_passive = await state_b.get_shared_secret_passive(shared_secret_active.header)
-    assert shared_secret_active.shared_secret   == shared_secret_passive.shared_secret
-    assert shared_secret_active.associated_data == shared_secret_passive.associated_data
+    shared_secret_active, associated_data_active, header = state_a.get_shared_secret_active(
+        get_bundle(state_b)
+    )
+    shared_secret_passive, associated_data_passive, _ = state_b.get_shared_secret_passive(header)
+    assert shared_secret_active == shared_secret_passive
+    assert associated_data_active == associated_data_passive
